requirements.txt
pyright
selenium
tinydb
deepdiff
python-dotenv